Create prototype filter sidebar.
This sidebar takes a dictionary of categories and subcategories (aka Tags), and creates a set of expandable accordions for each category with the tags associated within that category contained as buttons within the accordion body.

Buttons can be selected and, when the search button is clicked, an array of all tags is logged to the console.